+++ grelmicro/task/__init__.py
@@ -1,7 +1,7 @@
 """Task."""
 
+from grelmicro.task._tasks import Tasks
 from grelmicro.task.errors import TaskError
-from grelmicro.task.manager import TaskManager
 from grelmicro.task.router import TaskRouter
 
-__all__ = ["TaskError", "TaskManager", "TaskRouter"]
+__all__ = ["TaskError", "TaskRouter", "Tasks"]